feat: added create redirect api

+++ datanar/api/views.py
@@ -1,35 +1,90 @@
+import hashlib
+
 from django.views.generic import TemplateView
-from rest_framework import status, viewsets
+from rest_framework import status, viewsets, mixins
 from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 
 from redirects import models, serializers
+from sqids import Sqids
+from users import models as user_models
+
+
+sqids = Sqids()
 
 
 class APIDocumentationView(TemplateView):
     template_name = 'api/api_docs.html'
 
 
-class RedirectViewSet(viewsets.ViewSet):
+class RedirectViewSet(viewsets.ViewSet, mixins.CreateModelMixin):
+    def get_user(self, request):
+        if request.query_params.get("token"):
+            token = request.query_params.get("token")
+        else:
+            token = request.data.get("token")
+        user = user_models.User.objects.filter(username=token).first()
+        if request.user.is_authenticated:
+            user = request.user
+        return user
+
     def list(self, request):
-        queryset = models.Redirect.objects.all()
+        user = self.get_user(request)
+        if not user:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        queryset = models.Redirect.objects.filter(user=user)
         serializer = serializers.RedirectSerializer(queryset, many=True)
         return Response(serializer.data)
 
     def retrieve(self, request, pk=None):
+        user = self.get_user(request)
+        if not user:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         queryset = models.Redirect.objects.all()
-        redirect = get_object_or_404(queryset, pk=pk)
+        redirect = get_object_or_404(queryset, pk=pk, user=user)
         serializer = serializers.RedirectSerializer(redirect)
         return Response(serializer.data)
 
     def destroy(self, request, pk=None):
-        queryset = models.Redirect.objects.all()
-        redirect = get_object_or_404(queryset, pk=pk)
+        user = self.get_user(request)
+        if not user:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        queryset = models.Redirect.objects.filter(user=user)
+        redirect = get_object_or_404(queryset, pk=pk, user=user)
         redirect.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
-    def create(self, request):
-        pass
+    def create(self, request, *args, **kwargs):
+        if request.data.get("long_link") is None:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        serializer_data = {}
+
+        short_link = request.data.get("short_link")
+
+        if short_link and ("/" in short_link or models.Redirect.objects.get_by_short_link(short_link)):
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        if not short_link:
+            counter = 0
+            string = request.data.get("long_link")
+            while True:
+                temp_string = f"{string}{counter}" if counter > 0 else string
+                hash_object = hashlib.sha256(temp_string.encode())
+                number = int.from_bytes(hash_object.digest(), byteorder="big")
+                short_link = sqids.encode(list(map(int, list(str(number)))))[:5]
+                if not models.Redirect.objects.get_by_short_link(short_link):
+                    break
+                counter += 1
+
+        user = self.get_user(request)
+        if not user:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
+        serializer_data.update({"short_link": short_link, "user": user})
+        serializer = serializers.RedirectCreateSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        serializer.save(**serializer_data)
+        headers = self.get_success_headers(serializer.data)
+        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
 
 
 __all__ = [RedirectViewSet]
